Move tft_submodels into components
